pyproject.toml
[tool.poetry]
name = "NiimPrintX"
version = "0.1.0"
description = "App to print labels using NiimBot printer"
authors = ["Labbots <labbots@gmail.com>"]
license = "GNU GPLv3"
readme = "README.md"
packages = [{include = "*", from="NiimPrintX"}]


[tool.poetry.dependencies]
python = ">=3.12,<3.13"
click = "^8.1.7"
bleak = "^0.21.1"
loguru = "^0.7.2"
pillow = "^10.3.0"
pycairo = "^1.26.0"
wand = "^0.6.13"
appdirs = "^1.4.4"
rich = "^13.7.1"
bluepy = "^1.3.0"

[tool.poetry.group.dev.dependencies]
devtools = "^0.12.2"
pyinstaller = "^6.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
